Update docs on running plugins locally
These instructions referred to a file `autoresponder.js` which does not exist (a) for a plugin in general, or (b) for the autoresponder plugin in particular. The plugin template repo contains a file named `index.js` to export the plugin module, and this naming scheme seems to have been adopted widely (e.g. autoresponder, stale plugins both use `index.js`).